Update style.css
No idea if you are interested in this, but I made some purely style related changes to the theme to make it look better in my opinion. These changes make the site title more centered and evenly spaced with the rest of the content. It made more sense to me design-wise.