Update style.css
As a photographer, I would want my entry featured image to appear as it was meant to be displayed, not dark and washed out. Removing the dark background on the featured image looks nicer, and honestly makes more sense. The hover still functions, and I removed the annoying solid color flash by removing the z-index attribute.

With these changes, the featured image now acts just like the blog post images do. If you think people are going to want their header image to be dark, then I might suggest the option to disable the dark overlay. Either way, the z-index attribute was giving a strange solid color flash that ruined the smooth transition.